go/payment: rename files
